refactor: separate useGLTFComponent into its own file and update imports
