+++ packages/client-core/src/common/services/FileThumbnailJobState.tsx
@@ -73,7 +73,8 @@ import { useFind } from '@ir-engine/common'
 import config from '@ir-engine/common/src/config'
 import { getChildrenWithComponents } from '@ir-engine/ecs'
 import { uploadProjectFiles } from '@ir-engine/editor/src/functions/assetFunctions'
-import { useGLTFComponent, useTexture } from '@ir-engine/engine/src/assets/functions/resourceLoaderHooks'
+import { useTexture } from '@ir-engine/engine/src/assets/functions/resourceLoaderHooks'
+import { useGLTFComponent } from '@ir-engine/engine/src/assets/functions/useGLTFComponent'
 import { GLTFComponent } from '@ir-engine/engine/src/gltf/GLTFComponent'
 import { ErrorComponent } from '@ir-engine/engine/src/scene/components/ErrorComponent'
 import { ShadowComponent } from '@ir-engine/engine/src/scene/components/ShadowComponent'
