+++ packages/engine/src/assets/functions/resourceLoaderHooks.ts
@@ -26,16 +26,7 @@ Infinite Reality Engine. All Rights Reserved.
 import { useEffect } from 'react'
 import { AudioLoader, Texture } from 'three'
 
-import {
-  createEntity,
-  Entity,
-  entityExists,
-  getComponent,
-  removeEntity,
-  setComponent,
-  UndefinedEntity,
-  UUIDComponent
-} from '@ir-engine/ecs'
+import { Entity, UndefinedEntity } from '@ir-engine/ecs'
 import { NO_PROXY, useHookstate, useImmediateEffect } from '@ir-engine/hyperflux'
 import {
   Resource,
@@ -44,8 +35,6 @@ import {
   ResourceType
 } from '@ir-engine/spatial/src/resources/ResourceState'
 
-import { EntityTreeComponent } from '@ir-engine/ecs'
-import { GLTFComponent } from '../../gltf/GLTFComponent'
 import { FileLoader } from '../loaders/base/FileLoader'
 import { Loader } from '../loaders/base/Loader'
 import { parseStorageProviderURLs } from './parseSceneJSON'
@@ -156,42 +145,6 @@ async function getLoader<T extends ResourceAssetType>(
       loader
     )
   })
-}
-
-/**
- *
- * GLTF loader hook for use in React Contexts.
- * Creates an entity with a GLTFComponent as a child of the provided parentEntity param
- * Returns the root entity of the GLTF after the GLTF has completed loading
- *
- * @param url The URL of the GLTF file to load
- * @param parentEntity The entity that is loading the GLTF
- * @returns Entity | null
- */
-export function useGLTFComponent(url: string, parentEntity: Entity): Entity | null {
-  const gltfEntityState = useHookstate(UndefinedEntity)
-  const loaded = GLTFComponent.useSceneLoaded(gltfEntityState.value)
-
-  useEffect(() => {
-    if (!url || !parentEntity) return
-    const gltfEntity = createEntity()
-    setComponent(gltfEntity, EntityTreeComponent, { parentEntity })
-    setComponent(gltfEntity, UUIDComponent, {
-      entitySourceID: getComponent(parentEntity, UUIDComponent).entitySourceID,
-      entityID: UUIDComponent.generate()
-    })
-    setComponent(gltfEntity, GLTFComponent, { src: url })
-    gltfEntityState.set(gltfEntity)
-
-    return () => {
-      if (entityExists(gltfEntity)) {
-        removeEntity(gltfEntity)
-        gltfEntityState.set(UndefinedEntity)
-      }
-    }
-  }, [parentEntity, url])
-
-  return loaded ? gltfEntityState.value : null
 }
 
 /**
